refactor: split index.js code into app.js
